+++ src/controllers/image.js
@@ -1,0 +1,11 @@
+/** Los controladores son funciones que se pueden
+ * solicitar indicándolas como  argumento en las 
+ * funciones de enrutamiento.
+ * En este archivo se recopila todo lo que se puede 
+ * hacer con las imágenes
+ */
+
+const ctrl = {};
+
+
+module.exports = ctrl;