+++ src/routes/index.js
@@ -1,7 +1,11 @@
-module.exports = (app) => {
-  
-  // Página principal
-  app.get('/', (req, res) => {
-    res.send('Index page');
-  });
+const express = require('express');
+const router = express.Router();
+
+const home = require('../controllers/home' );
+
+
+module.exports = app => {
+    router.get('/', home.index);
+
+    app.use(router);
 };