+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v.1.0.1
+- Cambiado enrutamiento, usando objeto *Express.Router()*
+- Se agrega uso de controladores en el direccionamiento.
+
+
+
 ## v.1.0.0- Inicio del proyecto.
 - Estructura básica creada para proyecto de red social
 - modulos instalados y actualizados